pyproject.toml
[tool.poetry]
name = "nautobot-device-lifecycle-mgmt"
version = "1.2.0"
description = "Manages device lifecycles for platforms and software."
authors = ["Network to Code, LLC <opensource@networktocode.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/nautobot/nautobot-plugin-device-lifecycle-mgmt"
repository = "https://github.com/nautobot/nautobot-plugin-device-lifecycle-mgmt"
keywords = ["nautobot", "nautobot-plugin"]
include = [
    "LICENSE",
    "README.md",
]
packages = [
    { include = "nautobot_device_lifecycle_mgmt" },
]

[tool.poetry.dependencies]
python = "^3.7.1"
pycountry = "^22.3.5"
matplotlib = "^3.3.4"
nautobot = "^1.4.0"

[tool.poetry.extras]
openai = ["openai", "openVulnQuery"]

[tool.poetry.dev-dependencies]
invoke = "*"
black = "*"
bandit = "*"
pylint = "*"
pylint-django = "*"
pydocstyle = "*"
flake8 = "*"
pytest-django = "*"
sentry-sdk = "*"
django-debug-toolbar = "*"
coverage = "*"
django-extensions = "*"
time-machine = "*"
yamllint = "*"
mysqlclient = "*"
python-dotenv = "*"
openai = "*"
openVulnQuery = "*"
Markdown = "*"
# Rendering docs to HTML
mkdocs = "1.3.1"
# Material for MkDocs theme
mkdocs-material = "8.3.0"
# Render custom markdown for version added/changed/remove notes
mkdocs-version-annotations = "1.0.0"
# Automatic documentation from sources, for MkDocs
mkdocstrings = "0.19"
mkdocstrings-python = "0.7.1"

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | settings.py     # This is where you define files that should not be stylized by black
                     # the root of the project
)
'''

[tool.pylint.master]
# Include the pylint_django plugin to avoid spurious warnings about Django patterns
load-plugins="pylint_django"
ignore=["nautobot_device_lifecycle_mgmt/migrations"]

[tool.pylint.basic]
# No docstrings required for private methods (Pylint default), or for test_ functions, or for inner Meta classes.
no-docstring-rgx="^(_|test_|Meta$)"

[tool.pylint.messages_control]
# Line length is enforced by Black, so pylint doesn't need to check it.
# Pylint and Black disagree about how to format multi-line arrays; Black wins.
disable = """
    line-too-long,
    bad-continuation,
    duplicate-code,
    too-many-lines,
    too-many-ancestors,
    """

[tool.pylint.miscellaneous]
# Don't flag TODO as a failure, let us commit with things that still need to be done in the code
notes = """,
    FIXME,
    XXX,
    """

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
python_paths = "./"
testpaths = "tests/"
addopts = "-vv"
